moving SqlScopeOptions to external file
