+++ src/NServiceBus.SqlServer/NServiceBus.SqlServer.csproj
@@ -89,6 +89,7 @@
     <Compile Include="Addressing\QueueAddressParser.cs" />
     <Compile Include="Resharper.Annotations.g.cs" />
     <Compile Include="Sending\MessageDispatcher.cs" />
+    <Compile Include="Configuration\SqlScopeOptions.cs" />
     <Compile Include="SqlServerTransport.cs" />
     <Compile Include="Receiving\MessagePump.cs" />
     <Compile Include="Queuing\MessageReadResult.cs" />
